requirements.txt
munch~=2.5.0
timm~=0.4.12
PyYAML~=6.0

torchvision~=0.12.0
scikit-learn~=1.0.2
numpy~=1.21.5
matplotlib~=3.5.1
pillow~=9.0.1
fastapi~=0.78.0
pydantic~=1.9.0
yaml~=0.2.5